数禾促复借数据推客服 clc_usr_avl_lmt_lv0为double

+++ marketing-cs/src/main/java/com/br/marketing/speedconfig/MarketingCommonConfig.java
@@ -151,6 +151,6 @@ public class MarketingCommonConfig {
     /**
      * 促复借可用额度
      */
-    private Integer clcUsrAvlLmtLv0;
+    private Double clcUsrAvlLmtLv0;
 
 }
